Move diagnose endpoint out of health module
